Move ShadowProvider to org.robolectric.internal package.
