+++ robolectric-processor/src/test/java/org/robolectric/annotation/processing/validator/Utils.java
@@ -12,7 +12,7 @@ public class Utils {
   // "proper" directory. Otherwise the compile-testing suite is "smart" enough to find it even though
   // it hasn't been specified on the input file list.
   public static final JavaFileObject ROBO_SOURCE = forResource("mock-source/Robolectric.java");
-  public static final JavaFileObject SHADOW_PROVIDER_SOURCE = forResource("mock-source/org/robolectric/util/ShadowProvider.java");
+  public static final JavaFileObject SHADOW_PROVIDER_SOURCE = forResource("mock-source/org/robolectric/internal/ShadowProvider.java");
   public static final JavaFileObject SHADOW_EXTRACTOR_SOURCE = forResource("mock-source/org/robolectric/internal/ShadowExtractor.java");
 
   public static String toResourcePath(String clazzName) {
