+++ robolectric-processor/src/test/resources/META-INF/services/org.robolectric.internal.ShadowProvider
