+++ robolectric-processor/src/test/resources/mock-source/org/robolectric/internal/ShadowProvider.java
@@ -1,4 +1,4 @@
-package org.robolectric.util;
+package org.robolectric.internal;
 
 public interface ShadowProvider {
