--- robolectric-processor/src/test/resources/mock-source/org/robolectric/util/ShadowProvider.java
@@ -1,6 +1,0 @@
-package org.robolectric.util;
-
-public interface ShadowProvider {
-
-  void reset();
-}
