+++ robolectric-processor/src/test/resources/org/robolectric/Robolectric_Anything.java
@@ -7,7 +7,7 @@ import org.robolectric.annotation.processing.objects.Dummy;
 import org.robolectric.annotation.processing.shadows.ShadowAnything;
 import org.robolectric.annotation.processing.shadows.ShadowDummy;
 import org.robolectric.internal.ShadowExtractor;
-import org.robolectric.util.ShadowProvider;
+import org.robolectric.internal.ShadowProvider;
 
 @Generated("org.robolectric.annotation.processing.RobolectricProcessor")
 @SuppressWarnings({"unchecked","deprecation"})
