+++ robolectric-processor/src/test/resources/org/robolectric/Robolectric_NoExcludedTypes.java
@@ -3,7 +3,7 @@ package org.robolectric;
 import javax.annotation.Generated;
 
 import org.robolectric.internal.ShadowExtractor;
-import org.robolectric.util.ShadowProvider;
+import org.robolectric.internal.ShadowProvider;
 
 /**
  * Shadow mapper. Automatically generated by the Robolectric Annotation Processor.
