+++ robolectric-utils/src/main/java/org/robolectric/internal/ShadowProvider.java
@@ -1,0 +1,12 @@
+package org.robolectric.internal;
+
+/**
+ * Interface implemented by packages that provide shadows to Robolectric.
+ */
+public interface ShadowProvider {
+
+  /**
+   * Reset the static state of all shadows provided by this package.
+   */
+  void reset();
+}
